reggi/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "lxcor-reggi"
version = "0.1.0"
description = "API key management and user identity for Django — registration, email verification, key rotation, and tier assignment."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "lxcor", email = "i.alves.alex@gmail.com" }]
requires-python = ">=3.8"
keywords = ["django", "api-keys", "authentication", "registration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "django>=3.2",
    "djangorestframework>=3.14",
]

[project.urls]
Homepage = "https://github.com/lxcor/api-suite/tree/main/reggi"
Repository = "https://github.com/lxcor/api-suite"

[tool.setuptools.packages.find]
where = ["."]
include = ["reggi*"]
